% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{ascpR}
\alias{ascpR}
\title{Utility function to run aspera}
\usage{
ascpR(ascp_args, file, destDir = getwd())
}
\arguments{
\item{ascp_args}{Character vector of arguments to \code{ascp}.}

\item{file}{Url to aspera file to download.}

\item{destDir}{Path to directory to download \code{files} into.}
}
\value{
return code from call to ascp
}
\description{
Utility function to run aspera
}
\keyword{internal}
