% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{crawl_gse}
\alias{crawl_gse}
\title{Get GSE text from GEO}
\usage{
crawl_gse(gse_name)
}
\arguments{
\item{gse_name}{GEO study name to get metadata for}
}
\value{
Character vector of lines on GSE record.
}
\description{
Get GSE text from GEO
}
\examples{
gse_text <- crawl_gse('GSE111459')

}
