% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{crawl_gsms}
\alias{crawl_gsms}
\title{Crawls SRX pages for each GSM to get metadata.}
\usage{
crawl_gsms(gsm_names, max.workers = 50)
}
\arguments{
\item{gsm_names}{Character vector of GSMs.}

\item{max.workers}{Maximum number of parallel workers to split task betweem}
}
\value{
data.frame
}
\description{
Goes to each GSM page to get SRX then to each SRX page to get some more
metadata.
}
\examples{
srp_meta <- crawl_gsms("GSM3031462")

# returns NULL because records on dbGAP for privacy reasons
srp_meta <- crawl_gsms("GSM2439650")

# example with empty values
srp_meta <- crawl_gsms('GSM4043025')

}
