% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_characteristics.R
\name{esc}
\alias{esc}
\title{Preserve escape characters in grep patterns}
\usage{
esc(string)
}
\arguments{
\item{string}{Character vector with pattern}
}
\value{
\code{string} with escape characters preserved
}
\description{
Preserve escape characters in grep patterns
}
