% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{get_dldir}
\alias{get_dldir}
\title{Gets part of path to download bulk RNAseq sample from EBI or NCBI}
\usage{
get_dldir(srr, type = c("ebi", "ncbi"))
}
\arguments{
\item{srr}{SRR run name}

\item{type}{Either \code{'ebi'} or \code{'ncbi'}}
}
\value{
String path used by \code{\link{get_fastqs}}.
}
\description{
Gets part of path to download bulk RNAseq sample from EBI or NCBI
}
