% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{get_gsms}
\alias{get_gsms}
\title{Get GSMs needed to download RNA-seq data for a series}
\usage{
get_gsms(gse_name)
}
\arguments{
\item{gse_name}{GEO study name to get metadata for}
}
\value{
Character vector of sample GSMs for the series \code{gse_name}
}
\description{
Goes to a series GSE page to get sample GSMs.
}
\examples{

gsm_names <- get_gsms('GSE111459')

}
