% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{get_srp_meta}
\alias{get_srp_meta}
\title{Get metadata needed to download RNA-seq data for GSE}
\usage{
get_srp_meta(gse_name, data_dir = NULL)
}
\arguments{
\item{gse_name}{GEO study name to get metadata for}

\item{data_dir}{Path that folder with \code{gse_name} will be created in to save result. If \code{NULL} (default), results
will not be saved.}
}
\value{
\code{data.frame} with sample annotations for each GSE. Get's saved as \emph{data_dir/gse_name/gse_name.rds}.
}
\description{
Goes to GSE page to get GSMs then goes to each GSM page to get SRX then to each SRX page to get some more metadata.
}
\examples{

srp_meta <- get_srp_meta('GSE111459')
}
