% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{crawl_gsms}
\alias{crawl_gsms}
\title{Crawls SRX pages for each GSM to get metadata.}
\usage{
crawl_gsms(gsm_names, max.workers = 50)
}
\arguments{
\item{max.workers}{Maximum number of parallel workers to split task betweem}

\item{samples}{Character vector of GSMs.}
}
\value{
data.frame
}
\description{
Goes to each GSM page to get SRX then to each SRX page to get some more metadata.
}
\examples{
gsm_names <- get_gsms('GSE111459')
srp_meta <- crawl_gsms(gsm_names)

}
