% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_characteristics.R
\name{fix_characteristics}
\alias{fix_characteristics}
\title{Fix overlapping characteristics columns}
\usage{
fix_characteristics(pdata)
}
\arguments{
\item{pdata}{}
}
\value{

}
\description{
Fix overlapping characteristics columns
}
