% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp_utils.R
\name{get_fastqs}
\alias{get_fastqs}
\title{Download and fasterq-dump RNA-seq data from GEO}
\usage{
get_fastqs(srp_meta, data_dir, method = c("ftp", "aspera"))
}
\arguments{
\item{srp_meta}{\code{data.frame} with SRP meta info. Returned from \code{\link{get_srp_meta}}.}

\item{data_dir}{Path to folder that fastq files will be downloaded to. Will be created if doesn't exist.}

\item{method}{One of \code{'aspera'} or \code{'ftp'}. \code{'aspera'} is generally faster but requires the
ascp command line utility to be on your path. \code{'ftp'} is recommended if many fastqs are required as
a parallel for loop can give speeds comparable to \code{'aspera'}.}
}
\description{
First tries to get RNA-Seq fastq files from EBI. If not successful, gets SRA from GEO and converts to fastq.gz.
}
